ureadahead: Use upstream trace event mm_filemap_add_to_page_cache
Instead of using the trace events that are not upstream (and will likely
never be) that trace the opening of files, trace the portions of the file
that gets pulled into the page cache, which is exactly what ureadahead
wants.

The pros to this approach:

- Uses an upstream trace event and no longer requires a modified kernel
  that is different from upstream.

- Uses only a single trace event.

- No longer fails on relative paths.

- Shows only what is being pulled in when ureadahead is called, and not
  anything before hand (where it wastes time in doing so).

The trace event mm_filemap_add_to_page_cache records the following items:

  dev 254:3 ino 2b4d0 pfn=0x15b22a ofs=0 order=0

Where the device is: major 254 minor 3
The inode is: 0x2b4d0
The offset into the file is 0

With this information, the /proc/self/mountinfo is examined:

Which shows the device and file systems:

For the major:mino 254:3:

 27 1 254:3 / / rw,relatime shared:1 - ext4 /dev/vda3 rw,errors=remount-ro

We can see that this device is mounted at root '/'.

Then the file system is scanned (staying only on the file system, similar
to a "find . -xdev" walk, and will look for any file that has a matching
inode. Once that is found, that file can be used for the pack file. It
doesn't matter if there's more than one file with the same inode. Any one
will work with the readahead() system call for all the other files that
share the same inode.

Finally, the pages that were mapped in is examined and added to the pack
file for that file.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>