ureadahead: Use libtracefs and read the binary data directly
Instead of parsing the trace ASCII file, which is slow and error prone,
use the libtracefs interface that parses the kernel event format files
and can read the event binaries directly. This requires the 1.6 version
of libtracefs which CL 3916306 added.

This simplifies the code, and also allows for easier updates if the
events are to be changed later.

Also, use the macro FS_SYSTEM for the system file name. On some kernels,
the added open system call events are added in the trace system called,
"fs", and in others it's called "fs_trace". By having a macro that
defines this string, it is less intrusive to only have to change the
macro and not all the locations within it.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>