ureadahead: Consolidate reading of mountinfo
The reading of the trace will read /proc/mountinfo to map the device
numbers in the trace events to where they are mounted in the file system.
The writing of the pack file does the same as it looks at which device the
file it's about to write about, and finds its path too.

Cache the device path in the device hash and save a pointer to it with the
file, and remove the duplicate reading of /proc/mountinfo

Suggested-by: Gwendal Grignou <gwendal@chromium.org>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>