ureadahead: read trace's binary data directly
Read the binary trace file instead of parsing the text format of that,
which is faster and less error-prone. It does trace with the global
instance, but supporting another instance would be easy if necessary in
the future.

Also, this change defines FS_SYSTEM, which makes it less intrusive to
change "fs" subsystem when we want. On some kernels, the non-standard
open system call events are added in "fs", and in others it's called
"fs_trace".