ureadahead: Allow to work without syscall trace events
The do_sys_open and open_exec trace events were NACKED upstream and will
is not expected to ever be part of an upstream kernel. The new filemap
events have a chance of going upstream. If they do, allow ureadahead to
work with just them.

This requires doing a lookup of the /proc/self/mountinfo table to find the
mount points and search for the inodes that are extracted from the filemap
trace events to get the path names. Once they are found, those path names
are used to see what memory is present in the system with mincore. Note,
that these will soon not be needed as well.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>