ureadahead: fix memory corruption in add_inodes
add_inodes passes a region in inodes as a "parent" parameter to
trace_add_path, which is eventually used as a "parent" to nih's
allocation function. However, this parent requires to be a region
allocated by nih's allocation function family. So, add_inodes currently
violates this rule and that leads to meomry corruption. Fix it by
passing NULL as a parent, which is totally fine.

Signed-off-by: Takaya Saeki <takayas@chromium.org>