ureadahead: Follow a symlink during tracing phase
Current implemenetation of ureadahead filters out symlinks and others
that aren't "regular files" when looking into the tracing file.

This is handled by a single if statement that acts as a catch-all
to filter out unwanted file types. unfortunately, this means that any
data that are usually accessed via symlink (e.g. libraries) will not
be in pack file.

This patch will fix the issue by resolving the path of which symlink
points to instead of disregarding them.

Signed-off-by: Soichiro Fujii <soichirof@google.com>