remove NIH_BEGIN/END_EXTERN block from the headers
This is part of ongoing effort to decouple libnih from the ureadahead.

NIH_BEGIN/END_EXTERN expands to 'extern "C"' macro if `__cplusplus`
constant is defined.  this is only relevant if the ureadahead codebase
involves the usage of C++.

For now, remove them instead of replacing them with the
equivalent defined in the ureadahead code.

Signed-off-by: Soichiro Fujii <soichirof@google.com>