Fix: path name resolved from symlink can exceed PACK_PATH_MAX
Commit 3ee6736 ("ureadahead: Follow a symlink during tracing phase") has a
chance of storing invalid path name to a pack file, resulting in
unnecessary excess data.

That commit did not properly check if the length of the resolved path name
exceeds the length of PACK_PATH_MAX.  This path will be cut off into
length of PACK_PATH_MAX when storing into a pack file, turning it into an
invalid path that ureadahead won't be able to open during replay phase.

Fixes: 3ee6736 ("ureadahead: Follow a symlink during tracing phase")
Signed-off-by: Soichiro Fujii <soichirof@google.com>