Allow simultaneous recording and replaying
This change allows ureadahead to perform both recording
of the tracepoints and replaying the existing pack file
at the same time.

Currently, ureadahead attempts to read a pack file and
perform do_readahead if it loads successfully.
If it fails to locate a pack file, It will instead enable
kernel tracepoints and monitor the file access to generate
a new pack file.

Instead of only performing one step based on the existence
of pack file, this change allows ureadahead to simultaneously
record file access while replaying the pack file, by enabling
the tracepoints before the do_readahead step and generating
a new pack file after the readahead is done, which overwrites
the existing pack file.

Signed-off-by: Soichiro Fujii <soichirof@google.com>