Refactor flag-dependent control flow
`--dump` and `--force-trace` are a flag that acts more
like a command for a ureadahead, each changing how it
operates significantly. However, these flags are currently
handled in a single if statement block, fully coupled to
one another. This makes it difficult to understand how
code operates, and further complicates when trying to add
another feature backed by a flag.

Separate them into it's own control flow to make it easier
to understand, and give an error when both are set.

Signed-off-by: Soichiro Fujii <soichirof@google.com>