configure: Add touch to files to prevent rebuilding
Because git does not keep timestamps of files, running ./configure from a
git checkout can cause files that are already made to be remade and that
can make it fill he build because it may cause files to be built with
different versions of autoconf.

Add a touch command to update the timestamps of the files:

 Makefile.am Makefile.in aclocal.m4 configure config.h.in

This will prevent the ./configure from messing with the tree. It will even
not modify current files where git will show files were changed.

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>