Fix: missing close(fd) in trace code can leak open fd
Commit 61bd7e6 ("Use newly implemented set to replace NihHash")
replaced a usage of NihHash with newly introduced hash set.

One of the hash set usage is keeping track of files already
processed by the tracer. one check involves opening fd of a file,
obtaining device id & inode through it for checking if it's
already stored in the hash set. if it is, the file gets ignored.

While replacing the code that performs check mentioned above, it
accidentally removed `close(fd)` call that would have properly
closed the opened fd before moving onto the next file for
processing.

Re-introduce `close(fd)` to appropriate location so that fd
won't get leaked.

Fixes: 61bd7e6 ("Use newly implemented set to replace NihHash")
Signed-off-by: Soichiro Fujii <soichirof@google.com>